+++ cdp-agentkit-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Added `wow_create_token` action.
+
 ## [0.0.1] - 2024-11-04
 
 ### Added